Implement vendor product multi images feature

+++ app/Http/Controllers/Backend/VendorProductImageGalleryController.php
@@ -2,22 +2,26 @@
 
 namespace App\Http\Controllers\Backend;
 
-use App\DataTables\ProductImageGalleryDataTable;
+use App\DataTables\VendorProductImageGalleryDataTable;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Backend\CreateProductGalleryRequest;
 use App\Models\Product;
 use App\Models\ProductImageGallery;
 use App\Traits\ImageUploadTrait;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
-class ProductImageGalleryController extends Controller
+class VendorProductImageGalleryController extends Controller
 {
-
     use ImageUploadTrait;
-    public function index(Request $request, ProductImageGalleryDataTable $dataTable)
+
+    public function index(Request $request, VendorProductImageGalleryDataTable $dataTable)
     {
         $product = Product::findOrFail($request->product);
-        return $dataTable->render('admin.product.image-gallery.index', compact('product'));
+        if($product->vendor_id !== Auth::user()->vendor->id){
+            abort(404);
+        }
+        return $dataTable->render('vendor.product.image-gallery.index', compact('product'));
     }
 
 
@@ -34,6 +38,9 @@ class ProductImageGalleryController extends Controller
     public function destroy(string $id)
     {
         $gallery = ProductImageGallery::findOrFail($id);
+        if($gallery->product->vendor_id !== Auth::user()->vendor->id){
+            abort(404);
+        }
         $this->deleteImage($gallery->image);
         $gallery->delete();
